+++ 3-Greedy Algorithms, Minimum Spanning Tree, and Dynamic Programming/2-Dynamic Programming/2-Knapsack Problem/Knapsack.java
@@ -31,7 +31,7 @@ public class Knapsack {
      * a subproblem, we can cache its solution in a global take for O(1) lookup
      * time later on.
      */
-    private int[][] subproblems;
+    private double[][] subproblems;
 
     /**
      * Solves the knapsack problem of the items with the given values and
@@ -41,7 +41,7 @@ public class Knapsack {
      * @param cap capacity of the knapsack
      * @return included items
      */
-    public HashSet<Integer> knapsack(int[] vals, int[] weights, int cap) {
+    public HashSet<Integer> knapsack(double[] vals, int[] weights, int cap) {
         // Check whether the input arrays are null or empty
         if ((vals == null) || (vals.length == 0) || (weights == null) || (weights.length == 0)) {
             throw new IllegalArgumentException("The input values and weights should not be null or empty.");
@@ -53,7 +53,7 @@ public class Knapsack {
 
         int n = vals.length;
         // Initialization
-        subproblems = new int[n][cap + 1];
+        subproblems = new double[n][cap + 1];
         for (int x = 0; x <= cap; ++x) {
             if (weights[0] <= x) {
                 subproblems[0][x] = vals[0];
@@ -82,7 +82,7 @@ public class Knapsack {
      * @param cap capacity of the knapsack
      * @return included items
      */
-    private HashSet<Integer> reconstruct(int[] vals, int[] weights, int cap) {
+    private HashSet<Integer> reconstruct(double[] vals, int[] weights, int cap) {
         HashSet<Integer> includedItems = new HashSet<Integer>();
         int currItem = vals.length - 1, currCap = cap;
         while (currItem >= 1) {
